[#79] move toilet repository from application to domain
