+++ src/main/java/depromeet/lessonfour/server/toiletrecord/app/service/DeleteToiletRecordUseCase.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 import depromeet.lessonfour.server.common.annotation.UseCase;
 import depromeet.lessonfour.server.common.api.code.ErrorCode;
 import depromeet.lessonfour.server.common.exception.ServerException;
-import depromeet.lessonfour.server.toiletrecord.app.repository.ToiletRecordRepository;
 import depromeet.lessonfour.server.toiletrecord.domain.entity.ToiletRecord;
+import depromeet.lessonfour.server.toiletrecord.domain.repository.ToiletRecordRepository;
 import lombok.RequiredArgsConstructor;
 
 @UseCase
