+++ src/main/java/depromeet/lessonfour/server/toiletrecord/app/service/ToiletRecordQueryService.java
@@ -5,8 +5,10 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import depromeet.lessonfour.server.toiletrecord.app.repository.ToiletRecordRepository;
+import depromeet.lessonfour.server.common.domain.view.DailyExistenceView;
+import depromeet.lessonfour.server.common.domain.vo.ActivityAt;
 import depromeet.lessonfour.server.toiletrecord.domain.entity.ToiletRecord;
+import depromeet.lessonfour.server.toiletrecord.domain.repository.ToiletRecordRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -17,5 +19,14 @@ public class ToiletRecordQueryService {
 
   public List<ToiletRecord> findByDate(Long userId, LocalDate date) {
     return toiletRecordRepository.findByDate(userId, date);
+  }
+
+  public DailyExistenceView existsByActivityAt(Long userId, ActivityAt at) {
+    return toiletRecordRepository.existsByActivityAt(userId, at);
+  }
+
+  public List<DailyExistenceView> existsByActivityAt(
+      Long userId, ActivityAt start, ActivityAt end) {
+    return toiletRecordRepository.existsByActivityAt(userId, start, end);
   }
 }
